Decouple table sharing utils from organization contexts
