+++ spec/models/user_part_app_spec.rb
@@ -195,7 +195,6 @@ describe User do
 
   describe '#visualization_count' do
     include_context 'organization with users helper'
-    include TableSharing
 
     it 'filters by type if asked' do
       vis = FactoryGirl.create(:carto_visualization, user_id: @org_user_1.id, type: Carto::Visualization::TYPE_DERIVED)
