+++ spec/spec_helper_min.rb
@@ -5,6 +5,7 @@ require 'helpers/spec_helper_helpers'
 require 'helpers/named_maps_helper'
 require './spec/support/message_broker_stubs'
 require './spec/support/redis'
+require './spec/support/shared_entities_spec_helper'
 
 ENV['RAILS_ENV'] ||= 'test'
 # INFO: this is the only slow step of the test boot process
@@ -27,6 +28,7 @@ RSpec.configure do |config|
   config.include NamedMapsHelper
   config.include Capybara::DSL
   config.include FactoryGirl::Syntax::Methods
+  config.include SharedEntitiesSpecHelper
 
   config.after(:each) do
     Delorean.back_to_the_present
